Update and rename facecount.py to final.py
